+++ strand-braid-user/CHANGELOG.md
@@ -27,6 +27,10 @@
   was started into the newly created `cam_settings` directory inside the
   `.braidz` file.
 
+* Braid stores the feature detection settings for each camera determined with
+  Strand Camera was started into the newly created `feature_detect_settings`
+  directory inside the `.braidz` file.
+
 * MKV videos now save the camera name as metadata in the segment title field.
 
 * MKV, FMF, and uFMF videos have the camera name in the filename when saved with